+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 - Use the SQLite driver to write integration tests.
 - Write LevelDB Cordova plugin.
 
+## [0.6.9] 3-31-2017
+- Made `cleanArgs` method public.
+- Added primary key range optimizations for Level DB and Indexed DB.
+- Fixed issue with history when using UUIDs
+- History is now disabled for tables with no primary key.
+- Adjusted `BETWEEN` to be consistent regardless of data store being used.
+
 ## [0.6.8] 3-30-2017
 - Added selected table to Action/View filter function.
 - Added a misc data object to make passing around information easier.